+++ CHANGELOG.md
@@ -56,6 +56,7 @@ v1.8.0 [unreleased]
 - [#337](https://github.com/latera/homs/pull/337) Add claim button for widget page.
 - [#338](https://github.com/latera/homs/pull/338) Add task overview component.
 - [#336](https://github.com/latera/homs/pull/336) Update React to 16.10.2.
+- [#339](https://github.com/latera/homs/pull/339) Add polling with pagination for claiming table.
 
 v1.7.8 [2019-10-03]
 -------------------